put_strを分離
