+++ Makefile
@@ -6,7 +6,7 @@
 #    By: ttsubo <ttsubo@student.42.fr>              +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2025/03/09 12:57:47 by ttsubo            #+#    #+#              #
-#    Updated: 2025/03/17 13:39:11 by ttsubo           ###   ########.fr        #
+#    Updated: 2025/03/17 13:45:51 by ttsubo           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -26,7 +26,7 @@ LFT_DIR = lib/libft/
 MLX_DIR = lib/mlx/
 
 PRI_SRC = glx_key_input_pri.c glx_prv.c
-PUB_SRC = glx.c glx_color.c glx_img.c glx_key_input_pub.c
+PUB_SRC = glx.c glx_color.c glx_img.c glx_key_input_pub.c glx_str.c
 
 SRCS = $(addprefix $(GLX_PRI_DIR), $(PRI_SRC)) \
 	   $(addprefix $(GLX_PUB_DIR), $(PUB_SRC))
