+++ inc/glx.h
@@ -6,7 +6,7 @@
 /*   By: ttsubo <ttsubo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/09 11:47:22 by ttsubo            #+#    #+#             */
-/*   Updated: 2025/03/17 13:38:32 by ttsubo           ###   ########.fr       */
+/*   Updated: 2025/03/17 13:51:57 by ttsubo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,9 @@
 // custom includes
 # include "glx_color.h"
 # include "glx_define.h"
+# include "glx_img.h"
 # include "glx_private.h"
+# include "glx_str.h"
 # include "glx_struct.h"
 
 // submodule
@@ -25,50 +27,16 @@
 
 // standard library
 # include <X11/keysym.h>
-# include <stdbool.h>
-
-# define MAX_IMG 32
-# define KEY_MAX 0xFFFF
-# define UPDATE_COUNT 50000
-
-typedef struct s_glx_prv
-{
-	size_t		update_count;
-	
-	void		(*error)(t_glx *, char *);
-}				t_glx_prv;
-
-typedef struct s_glx
-{
-	t_glx_prv	*_;
-	size_t		frame_count;
-	void		*mlx;
-	void		*win;
-	void		*imgs[MAX_IMG];
-	int			imgc;
-	bool		key_state[KEY_MAX];
-	bool		key_just_state[KEY_MAX];
-	int			(*btnp)(t_glx *, int);
-	int			(*update)(t_glx *);
-	int			(*draw)(t_glx *);
-	void		(*load_img)(t_glx *, char *, int, int);
-	void		(*put_img)(t_glx *, int, t_pos);
-	void		(*put_str)(t_glx *, char *, t_pos, t_glx_color_i);
-	void		(*run)(t_glx *, int (*)(t_glx *), int (*)(t_glx *));
-	void		(*quit)(t_glx *, int);
-}				t_glx;
 
 // private
-void			glx_error(t_glx *self, char *mes);
+void	glx_error(t_glx *self, char *mes);
 
 // public
-t_glx			*glx_init(char *title, int win_w, int win_h);
-void			glx_load_img(t_glx *self, char *path, int w, int h);
-void			glx_put_img(t_glx *self, int img_i, t_pos pos);
-void			glx_put_str(t_glx *self, char *str, t_pos pos,
-					t_glx_color_i color);
-void			glx_run(t_glx *self, int (*update)(t_glx *),
-					int (*draw)(t_glx *));
-void			glx_quit(t_glx *self, int sts_code);
-int				glx_btnp(t_glx *self, int keycode);
+t_glx	*glx_init(char *title, int win_w, int win_h);
+void	glx_load_img(t_glx *self, char *path, int w, int h);
+void	glx_put_img(t_glx *self, int img_i, t_pos pos);
+void	glx_put_str(t_glx *self, char *str, t_pos pos, t_glx_color_i color);
+void	glx_run(t_glx *self, int (*update)(t_glx *), int (*draw)(t_glx *));
+void	glx_quit(t_glx *self, int sts_code);
+int		glx_btnp(t_glx *self, int keycode);
 #endif
